backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.3.4  # Settings management

# Database
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.7

# Configuration
python-dotenv==1.0.1

# AI/ML
onnxruntime==1.18.1
transformers==4.44.2
numpy==1.26.4
torch==2.1.0  # If using PyTorch models
# tensorflow==2.15.0  # Alternative: if using TensorFlow

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.9  # Form data parsing
email-validator==2.1.0  # Email validation

# API Documentation & Testing
httpx==0.25.2  # Async HTTP client for testing
pytest==7.4.3  # Testing framework
pytest-asyncio==0.21.1  # Async test support
pytest-cov==4.1.0  # Coverage reporting

# Utility Libraries
python-dateutil==2.8.2  # Date/time manipulation
pytz==2024.1  # Timezone support
aiofiles==23.2.1  # Async file operations
celery==5.3.4  # Task queue (if needed)
# flower==2.0.1  # Celery monitoring (optional)

# Monitoring & Logging
loguru==0.7.2  # Better logging
sentry-sdk[fastapi]==1.40.0  # Error tracking (optional)
prometheus-client==0.19.0  # Metrics (optional)

# Development Tools
black==23.12.1  # Code formatter
flake8==6.1.0  # Linter
mypy==1.8.0  # Type checking
ipython==8.18.1  # Better Python shell
pre-commit==3.6.0  # Git hooks

# Data Processing
pandas==2.1.4  # Data manipulation (if needed)
openpyxl==3.1.2  # Excel file support
python-docx==1.1.0  # Word document support
PyPDF2==3.0.1  # PDF processing
Pillow==10.2.0  # Image processing

# API Integration
httpx==0.25.2  # Modern HTTP client
requests==2.31.0  # Simple HTTP library
websockets==12.0  # WebSocket support
